mostly finished implementation of capabilities-related endpoints
