+++ config.yml
@@ -1,3 +1,6 @@
 api_version: 1.1.0
-stac_version: 1.0.0
+stac_version: 0.9.0
 url: http://www.brockmann-consult.de/xcube-geoDB-openEO
+id: xcube-geodb-openeo
+title: xcube geoDB Server, openEO API
+description: Catalog of geoDB collections.
